+++ libbeat/docs/processors-using.asciidoc
@@ -43,7 +43,7 @@ The supported processors are:
  * <<drop-event,`drop_event`>>
  * <<drop-fields,`drop_fields`>>
  * <<include-fields,`include_fields`>>
- * <<kubernetes,`kubernetes`>>
+ * <<add-kubernetes-metadata,`add_kubernetes_metadata`>>
  * <<add-docker-metadata,`add_docker_metadata`>>
 
 [[conditions]]
@@ -492,17 +492,18 @@ section.
 NOTE: If you define an empty list of fields under `include_fields`, then only
 the required fields, `@timestamp` and `type`, are exported.
 
-[[kubernetes]]
+[[add-kubernetes-metadata]]
 === Adding Kubernetes Metadata
 
-The `kubernetes` processor annotates each event with relevant metadata based on
-which Kubernetes pod the event originated from. Each event is annotated with:
+The `add_kubernetes_metadata` processor annotates each event with relevant
+metadata based on which Kubernetes pod the event originated from. Each event is
+annotated with:
 
 * Pod Name
 * Namespace
 * Labels
 
-The `kubernetes` processor has two basic building blocks which are:
+The `add_kubernetes_metadata` processor has two basic building blocks which are:
 
 * Indexers
 * Matchers
@@ -526,7 +527,7 @@ Kubernetes.
 [source,yaml]
 -------------------------------------------------------------------------------
 processors:
-- kubernetes:
+- add_kubernetes_metadata:
     in_cluster: true
 -------------------------------------------------------------------------------
 
@@ -536,7 +537,7 @@ the Kubernetes node.
 [source,yaml]
 -------------------------------------------------------------------------------
 processors:
-- kubernetes:
+- add_kubernetes_metadata:
     in_cluster: false
     host: <hostname>
     kube_config: ${HOME}/.kube/config
@@ -548,7 +549,7 @@ enables ones that the user is interested in.
 [source,yaml]
 -------------------------------------------------------------------------------
 processors:
-- kubernetes:
+- add_kubernetes_metadata:
     in_cluster: false
     host: <hostname>
     kube_config: ~/.kube/config
