+++ libbeat/processors/add_kubernetes_metadata/registry.go
@@ -1,4 +1,4 @@
-package kubernetes
+package add_kubernetes_metadata
 
 import (
 	"errors"
